Rename package cap to ocap, to avoid colliding with builtin 'cap'.
