+++ pkg/client/client.go
@@ -10,9 +10,9 @@ import (
 	"capnproto.org/go/capnp/v3/rpc"
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/peer"
-	"github.com/wetware/ww/pkg/cap/anchor"
-	"github.com/wetware/ww/pkg/cap/cluster"
-	"github.com/wetware/ww/pkg/cap/pubsub"
+	"github.com/wetware/ww/pkg/ocap/anchor"
+	"github.com/wetware/ww/pkg/ocap/cluster"
+	"github.com/wetware/ww/pkg/ocap/pubsub"
 	"github.com/wetware/ww/pkg/vat"
 )
