+++ pkg/client/host.go
@@ -9,8 +9,8 @@ import (
 	"capnproto.org/go/capnp/v3/rpc"
 	"github.com/libp2p/go-libp2p-core/peer"
 	ma "github.com/multiformats/go-multiaddr"
-	"github.com/wetware/ww/pkg/cap/anchor"
-	"github.com/wetware/ww/pkg/cap/cluster"
+	"github.com/wetware/ww/pkg/ocap/anchor"
+	"github.com/wetware/ww/pkg/ocap/cluster"
 	"github.com/wetware/ww/pkg/vat"
 )
