+++ pkg/ocap/anchor/path_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/wetware/ww/pkg/cap/anchor"
+	"github.com/wetware/ww/pkg/ocap/anchor"
 )
 
 func TestPath(t *testing.T) {
