+++ pkg/ocap/anchor/stm_test.go
@@ -6,7 +6,7 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/wetware/ww/pkg/cap/anchor"
+	"github.com/wetware/ww/pkg/ocap/anchor"
 )
 
 func TestNewScheduler(t *testing.T) {
