+++ pkg/ocap/doc.go
@@ -1,0 +1,2 @@
+// Package ocap contains implementations for the object capabilities defined in api/
+package ocap
