+++ pkg/ocap/pubsub/subscription_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/require"
 	chan_api "github.com/wetware/ww/internal/api/channel"
 	api "github.com/wetware/ww/internal/api/pubsub"
-	"github.com/wetware/ww/pkg/cap/channel"
+	"github.com/wetware/ww/pkg/ocap/channel"
 )
 
 func TestSubscription_refcount(t *testing.T) {
